refactor: 공통 컴포넌트 shared/ui로 이동 및 forwardRef 제거
