+++ src/shared/ui/Card.tsx
@@ -1,0 +1,39 @@
+interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+  ref?: React.Ref<HTMLDivElement>;
+}
+
+interface CardHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+  ref?: React.Ref<HTMLDivElement>;
+}
+
+interface CardTitleProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+  ref?: React.Ref<HTMLDivElement>;
+}
+
+interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+  ref?: React.Ref<HTMLDivElement>;
+}
+
+export const Card = ({ className, ref, ...props }: CardProps) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+)
+Card.displayName = "Card"
+
+export const CardHeader = ({ className, ref, ...props }: CardHeaderProps) => (
+  <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+)
+CardHeader.displayName = "CardHeader"
+
+export const CardTitle = ({ className, ref, ...props }: CardTitleProps) => (
+  <h3 ref={ref} className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
+)
+CardTitle.displayName = "CardTitle"
+
+export const CardContent = ({ className, ref, ...props }: CardContentProps) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+)
+CardContent.displayName = "CardContent"
